feat: add confirmation on delete user and promotion

+++ front/public/scripts/promotion-scripts.js
@@ -49,37 +49,42 @@ registerForm.addEventListener("submit", function (e) {
 
 // delete promotion
 async function deletePromotion(promotionId) {
+  const aux = confirm(`Deseja deletar a promoção ${promotionId}?`);
+
+  if(!aux){
+    return;
+  }
   fetch(`/promotion/${promotionId}`, {
     method: "DELETE",
   })
-    .then((res) => {
-      if (res.status === 200 || res.status === 204) {
-        document.getElementById(`promotion_${promotionId}`).remove();
-        iziToast.success({
-          title: "Sucesso",
-          message: "Promoção deletada com sucesso!",
-          position: "bottomLeft",
-        });
-      } else {
-        res
-          .json()
-          .then((data) => {
-            iziToast.info({
-              title: "Info",
-              message: data.message,
-              position: "bottomLeft",
-            });
-          })
-          .catch((err) => {
-            throw new Error();
-          });
-      }
-    })
-    .catch((_) => {
-      iziToast.error({
-        title: "Erro",
-        message: "Erro ao deletar promoção!",
+  .then((res) => {
+    if (res.status === 200 || res.status === 204) {
+      document.getElementById(`promotion_${promotionId}`).remove();
+      iziToast.success({
+        title: "Sucesso",
+        message: "Promoção deletada com sucesso!",
         position: "bottomLeft",
       });
+    } else {
+      res
+        .json()
+        .then((data) => {
+          iziToast.info({
+            title: "Info",
+            message: data.message,
+            position: "bottomLeft",
+          });
+        })
+        .catch((err) => {
+          throw new Error();
+        });
+    }
+  })
+  .catch((_) => {
+    iziToast.error({
+      title: "Erro",
+      message: "Erro ao deletar promoção!",
+      position: "bottomLeft",
     });
+  });
 }
